feat: custom exceptions setup

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -16,6 +17,8 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var deleteLeaveType = await _leaveTypeRepository.GetByIdAsync(request.Id);
+        if (deleteLeaveType == null)
+            throw new NotFoundException(nameof(LeaveType), request.Id);
         
         await _leaveTypeRepository.DeleteAsync(deleteLeaveType);
